Split pandoc splice modules
