+++ README.md
@@ -83,7 +83,7 @@ Before tests (tasks impacting the larger architectural context in code base),
 
 Milestone (notes.srid.ca),
 
-- [ ] Footnotes
+- [x] Footnotes
 - [x] Custom route slugs https://github.com/srid/emanote/discussions/42
 - Blog post friendly
   - Queries
