+++ src/Emanote/Model/Type.hs
